Kreatoroom/requirements.txt
Flask-SQLAlchemy==2.4.1
SQLAlchemy==1.3.15
PyYAML==5.3.1
Flask==1.1.1
pandas==1.0.3
google-cloud-storage==1.42.3
google-cloud-bigquery==2.28.1
google-api-python-client==2.26.1
pymysql==0.9.3
scikit-learn==0.21.2
numpy==1.17.2
seaborn==0.9.0
matplotlib==3.1.3
pytest==5.3.5
mlxtend==0.17.2

google-cloud-storage
google-cloud-sql-connector